requirements.txt
Django==4.2
django-cities-light==3.9.2
django-countries==7.5.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.5.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
